pyproject.toml
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "devworld"
version = "0.0.0"
description = "Demo for the talk about AI-agents at DevWorld 2025"
authors = [
    "Alex Shershebnev <alexs@forgood.ai>",
]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.12"

anthropic = "^0.47.0"
beautifulsoup4 = "^4.13.3"
googlesearch-python = "^1.3.0"
langchain_anthropic = "^0.3.8"
langchain_community = "^0.3.18"
langchain-experimental = "^0.3.4"
langchain-mcp-adapters = "^0.0.2"
langgraph-cli = {version = "^0.1.73", extras = ["inmem"]}
lxml = "5.3.1"
mcp = "^1.3.0"
mcp-server-fetch = "^2025.1.17"
notebook = "^7.2.2"
openai = "^1.65.0"
pydantic = "^2.10.6"
python-dotenv = "^1.0.1"
requests = "^2.32.3"